requirements/common.txt
Django==1.6
pyshp==1.2.0
requests==2.3.0
django-tastypie==0.11.1


